create notification module for server and move logic
