+++ src/main/java/de/tum/cit/aet/application/service/ApplicationService.java
@@ -5,7 +5,7 @@ import de.tum.cit.aet.application.domain.Application;
 import de.tum.cit.aet.application.domain.dto.*;
 import de.tum.cit.aet.application.repository.ApplicationRepository;
 import de.tum.cit.aet.core.constants.DocumentType;
-import de.tum.cit.aet.core.constants.EmailType;
+import de.tum.cit.aet.notification.constants.EmailType;
 import de.tum.cit.aet.core.constants.Language;
 import de.tum.cit.aet.core.domain.Document;
 import de.tum.cit.aet.core.domain.DocumentDictionary;
@@ -14,8 +14,8 @@ import de.tum.cit.aet.core.exception.OperationNotAllowedException;
 import de.tum.cit.aet.core.service.CurrentUserService;
 import de.tum.cit.aet.core.service.DocumentDictionaryService;
 import de.tum.cit.aet.core.service.DocumentService;
-import de.tum.cit.aet.core.service.EmailService;
-import de.tum.cit.aet.core.service.mail.Email;
+import de.tum.cit.aet.notification.service.EmailService;
+import de.tum.cit.aet.notification.service.mail.Email;
 import de.tum.cit.aet.job.domain.Job;
 import de.tum.cit.aet.job.repository.JobRepository;
 import de.tum.cit.aet.usermanagement.constants.GradingScale;
