+++ src/main/java/de/tum/cit/aet/notification/constants/TemplateVariable.java
@@ -1,4 +1,4 @@
-package de.tum.cit.aet.core.constants;
+package de.tum.cit.aet.notification.constants;
 
 import java.util.Arrays;
 import java.util.Set;
