+++ src/main/java/de/tum/cit/aet/notification/dto/EmailTemplateOverviewDTO.java
@@ -1,6 +1,6 @@
-package de.tum.cit.aet.core.dto;
+package de.tum.cit.aet.notification.dto;
 
-import de.tum.cit.aet.core.constants.EmailType;
+import de.tum.cit.aet.notification.constants.EmailType;
 import java.util.UUID;
 
 /**
