+++ src/main/java/de/tum/cit/aet/notification/dto/EmailTemplateTranslationDTO.java
@@ -1,3 +1,3 @@
-package de.tum.cit.aet.core.dto;
+package de.tum.cit.aet.notification.dto;
 
 public record EmailTemplateTranslationDTO(String subject, String body) {}
