+++ src/main/java/de/tum/cit/aet/notification/repository/EmailTemplateRepository.java
@@ -1,12 +1,10 @@
-package de.tum.cit.aet.core.repository;
+package de.tum.cit.aet.notification.repository;
 
-import de.tum.cit.aet.core.constants.EmailType;
-import de.tum.cit.aet.core.domain.EmailTemplate;
-import de.tum.cit.aet.core.dto.EmailTemplateOverviewDTO;
+import de.tum.cit.aet.core.repository.TumApplyJpaRepository;
+import de.tum.cit.aet.notification.constants.EmailType;
+import de.tum.cit.aet.notification.domain.EmailTemplate;
+import de.tum.cit.aet.notification.dto.EmailTemplateOverviewDTO;
 import de.tum.cit.aet.usermanagement.domain.ResearchGroup;
-import java.util.Optional;
-import java.util.Set;
-import java.util.UUID;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -14,6 +12,10 @@ import org.springframework.data.jpa.repository.EntityGraph;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+import java.util.Set;
+import java.util.UUID;
 
 @Repository
 public interface EmailTemplateRepository extends TumApplyJpaRepository<EmailTemplate, UUID> {
@@ -59,7 +61,7 @@ public interface EmailTemplateRepository extends TumApplyJpaRepository<EmailTemp
      */
     @Query(
         """
-            SELECT new de.tum.cit.aet.core.dto.EmailTemplateOverviewDTO(
+            SELECT new de.tum.cit.aet.notification.dto.EmailTemplateOverviewDTO(
                 et.emailTemplateId,
                 et.templateName,
                 et.emailType,
