+++ src/main/java/de/tum/cit/aet/notification/service/EmailTemplateService.java
@@ -1,27 +1,27 @@
-package de.tum.cit.aet.core.service;
+package de.tum.cit.aet.notification.service;
 
-import de.tum.cit.aet.core.constants.EmailType;
 import de.tum.cit.aet.core.constants.Language;
-import de.tum.cit.aet.core.domain.EmailTemplate;
-import de.tum.cit.aet.core.domain.EmailTemplateTranslation;
-import de.tum.cit.aet.core.domain.EmailTemplate_;
-import de.tum.cit.aet.core.dto.*;
+import de.tum.cit.aet.core.dto.PageDTO;
+import de.tum.cit.aet.core.dto.PageResponseDTO;
 import de.tum.cit.aet.core.exception.EmailTemplateException;
 import de.tum.cit.aet.core.exception.EntityNotFoundException;
 import de.tum.cit.aet.core.exception.ResourceAlreadyExistsException;
 import de.tum.cit.aet.core.exception.TemplateProcessingException;
-import de.tum.cit.aet.core.repository.EmailTemplateRepository;
+import de.tum.cit.aet.core.service.CurrentUserService;
 import de.tum.cit.aet.core.util.HtmlSanitizer;
 import de.tum.cit.aet.core.util.TemplateUtil;
 import de.tum.cit.aet.evaluation.constants.RejectReason;
+import de.tum.cit.aet.notification.constants.EmailType;
+import de.tum.cit.aet.notification.domain.EmailTemplate;
+import de.tum.cit.aet.notification.domain.EmailTemplateTranslation;
+import de.tum.cit.aet.notification.domain.EmailTemplate_;
+import de.tum.cit.aet.notification.dto.EmailTemplateDTO;
+import de.tum.cit.aet.notification.dto.EmailTemplateOverviewDTO;
+import de.tum.cit.aet.notification.dto.EmailTemplateTranslationDTO;
+import de.tum.cit.aet.notification.repository.EmailTemplateRepository;
 import de.tum.cit.aet.usermanagement.domain.ResearchGroup;
 import de.tum.cit.aet.usermanagement.domain.User;
 import jakarta.transaction.Transactional;
-import java.io.IOException;
-import java.util.HashSet;
-import java.util.Objects;
-import java.util.Set;
-import java.util.UUID;
 import lombok.AllArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
@@ -29,6 +29,12 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
+
+import java.io.IOException;
+import java.util.HashSet;
+import java.util.Objects;
+import java.util.Set;
+import java.util.UUID;
 
 @Service
 @AllArgsConstructor
