+++ src/main/java/de/tum/cit/aet/notification/web/EmailSettingResource.java
@@ -1,8 +1,8 @@
-package de.tum.cit.aet.core.web;
+package de.tum.cit.aet.notification.web;
 
-import de.tum.cit.aet.core.dto.EmailSettingDTO;
+import de.tum.cit.aet.notification.dto.EmailSettingDTO;
 import de.tum.cit.aet.core.service.CurrentUserService;
-import de.tum.cit.aet.core.service.EmailSettingService;
+import de.tum.cit.aet.notification.service.EmailSettingService;
 import java.util.Set;
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
