+++ src/main/java/de/tum/cit/aet/usermanagement/domain/User.java
@@ -1,7 +1,7 @@
 package de.tum.cit.aet.usermanagement.domain;
 
 import de.tum.cit.aet.core.domain.AbstractAuditingEntity;
-import de.tum.cit.aet.core.domain.EmailSetting;
+import de.tum.cit.aet.notification.domain.EmailSetting;
 import de.tum.cit.aet.job.domain.Job;
 import jakarta.persistence.*;
 import java.time.LocalDate;
